requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async and Concurrency
aioredis==2.0.1
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Machine Learning and GNN
torch==2.1.1
torch-geometric==2.4.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-cluster==1.6.3
torch-spline-conv==1.2.2
numpy==1.24.3
scipy==1.11.4

# Evolutionary Algorithms
deap==1.4.1
pymoo==0.6.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Configuration and Logging
pyyaml==6.0.1
structlog==23.2.0
python-json-logger==2.0.7

# Monitoring and Metrics
prometheus-client==0.19.0
grafana-api==1.0.3

# Cloud and Storage
boto3==1.34.0
minio==7.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Kubernetes
kubernetes==28.1.0

# Additional Utilities
python-multipart==0.0.6
httpx==0.25.2
websockets==12.0 